// packages/bridge-ui-next/src/config/generated/chainConfig.ts
// Generated by generateChainConfig on 5/31/2026, 12:20:44 AM
// WARNING: Do not change this file manually as it will be overwritten
import type { ChainConfigMap } from "$libs/chain";

enum LayerType {
  L1 = "L1",
  L2 = "L2",
  L3 = "L3",
}

export const chainConfig: ChainConfigMap = {
  17000: {
    name: "Holesky",
    type: LayerType.L1,
    icon: "eth",
    rpcUrls: {
      default: { http: ["https://ethereum-holesky-rpc.publicnode.com"] },
    },
    nativeCurrency: { name: "Ether", symbol: "ETH", decimals: 18 },
    blockExplorers: {
      default: { name: "Etherscan", url: "https://holesky.etherscan.io" },
    },
  },
  167009: {
    name: "Taiko Hekla",
    type: LayerType.L2,
    icon: "taiko",
    rpcUrls: { default: { http: ["https://rpc.hekla.taiko.xyz"] } },
    nativeCurrency: { name: "Ether", symbol: "ETH", decimals: 18 },
    blockExplorers: {
      default: { name: "Taikoscan", url: "https://hekla.taikoscan.network" },
    },
  },
};